Make sure we deploy to the custom domain
We've got GitHub Actions set up to build and deploy the site for us but
it keeps deploying to https://bridgefoundry.github.io/operations instead
of https://operations.bridgefoundry.org.

The reason for this is that the CNAME does not get included as part of
the files that get generated in the `gh-pages` branch. In order to
include it, we have to move the file to the `docs` directory.